Don't auto-start services on installation or upgrades
The default configuration for keylime agent specifies a
default TPM password (“keylime”). If the agent initializes
the TPM device with the above password, it later becomes
very difficult to change this.

Starting the keylime agent atop the default configuration
(python3-keylime-defaultconfig) is unlikely to be successful
anyway, since the default configuration assumes a co-located
keylime verifier/registrar; very unlikely in a production
setting.

Closes: #29